feat(frontend): rename Grid to Game
