+++ frontend/src/components/Game.svelte
